Extract e2e misc helpers
